+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.8.0
+  * Streams the credentials cannot access (403) are now excluded from the catalog during discovery instead of raising an error
+  * Added unit tests for stream access check during discovery
+
 ## 1.7.3
   * Bump requests library version [#71](https://github.com/singer-io/tap-mixpanel/pull/71)